+++ tests/e2e/cases/date_instances/in.ts
@@ -1,3 +1,5 @@
+process.env.TZ = "UTC";
+
 const fixed = new Date(1234);
 const later = new Date(fixed.valueOf() + 66);
 const now = new Date();